+++ src/providers/mod.rs
@@ -1,7 +1,9 @@
 mod completion;
 mod definition;
 mod document_link;
+mod hover;
 
-pub use completion::complete;
+pub use completion::provide_completion;
 pub use definition::provide_definition;
 pub use document_link::provide_document_links;
+pub use hover::provide_hover;